Forgot that this is meant to remain py3.8 compatible. whoopsie!
this is because i'm still running an old mac mini that maxes out at
macos 10.15

incidentally i'm removing the type annotations from prompt.py because
it's too annoying to make it play nice in 3.8. it's not as simple as
using TYPE_CHECKING because Callable exists in 3.8 but is more
restrictive on what parameters it can take (i.e. no ParamSpec)

at least it's something like that i cbf figuring it out fully